+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 - New: Add `--skip-duplicate` to NuGet push to skip pushing packages that already exist in
   order to make it easier to retry released that have failed partially
+- New: Make it possible to disable Helm chart linting as well as enable strict linting
+- Fix: Update Helm charts before linting them
 - Fix: Properly set version when publishing `.csproj` projects
 
 # 0.31-beta